Refactor views to new viewmodels
